requirements.txt
fastapi
uvicorn
python-dotenv
httpx
openai>=1.0.0
